refactor(adapter-server): extract data-quality route to ai-data-quality.ts
Move POST /api/ai/data-quality handler into its own module to address
CodeRabbit file-size concern (ai-api.ts was growing beyond readable size).
No logic changes — same validation, same dataProvider.query call, same
scanDataQuality invocation.
